+++ test/glpk_tests.jl
@@ -21,6 +21,7 @@ if @isdefined GLPK
 
     @testset "Solving storage_test with GLPK" begin
         dbfile = joinpath(@__DIR__, "storage_test.sqlite")
+        #dbfile = "c:/temp/storage_test.sqlite"
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         # Test with default outputs
@@ -151,6 +152,15 @@ if @isdefined GLPK
         @test isapprox(testqry[1,:val], 3840.94023817782; atol=TOL)
         @test isapprox(testqry[2,:val], 3427.81584479179; atol=TOL)
 
+        # Test MinimumUtilization
+        SQLite.DBInterface.execute(db, "insert into MinimumUtilization select ROWID, '1', 'gas', val, 2025, 0.5 from TIMESLICE")
+        NemoMod.calculatescenario(dbfile; jumpmodel = Model(optimizer_with_attributes(GLPK.Optimizer, "presolve" => true)), numprocs=1, varstosave="vproductionbytechnologyannual")
+        testqry = SQLite.DBInterface.execute(db, "select * from vproductionbytechnologyannual where t = 'gas' and y = 2025") |> DataFrame
+
+        @test isapprox(testqry[1,:val], 15.768; atol=TOL)
+
+        SQLite.DBInterface.execute(db, "delete from MinimumUtilization")
+
         # Delete test results and re-compact test database
         NemoMod.dropresulttables(db)
         testqry = SQLite.DBInterface.execute(db, "VACUUM")
@@ -158,6 +168,7 @@ if @isdefined GLPK
 
     @testset "Solving storage_transmission_test with GLPK" begin
         dbfile = joinpath(@__DIR__, "storage_transmission_test.sqlite")
+        #dbfile = "c:/temp/storage_transmission_test.sqlite"
         chmod(dbfile, 0o777)  # Make dbfile read-write. Necessary because after Julia 1.0, Pkg.add makes all package files read-only
 
         NemoMod.calculatescenario(dbfile; jumpmodel=Model(optimizer_with_attributes(GLPK.Optimizer, "presolve" => true)),
@@ -189,6 +200,15 @@ if @isdefined GLPK
         @test isapprox(testqry[8,:val], 178.714990000842; atol=TOL)
         @test isapprox(testqry[9,:val], 170.204752381754; atol=TOL)
         @test isapprox(testqry[10,:val], 162.099764173099; atol=TOL)
+
+        # Test MinimumUtilization
+        SQLite.DBInterface.execute(db, "insert into MinimumUtilization select ROWID, '1', 'gas', val, 2025, 0.2 from TIMESLICE")
+        NemoMod.calculatescenario(dbfile; jumpmodel = jumpmodel=Model(optimizer_with_attributes(GLPK.Optimizer, "presolve" => true)), numprocs=1, varstosave="vproductionbytechnologyannual", calcyears=[2020,2025,2029])
+        testqry = SQLite.DBInterface.execute(db, "select * from vproductionbytechnologyannual where t = 'gas' and y = 2025") |> DataFrame
+
+        @test isapprox(testqry[1,:val], 16.3149963697108; atol=TOL)
+
+        SQLite.DBInterface.execute(db, "delete from MinimumUtilization")
 
         # Delete test results and re-compact test database
         NemoMod.dropresulttables(db)